refactor UserMapper package; update import statements in service and test classes
